Merged PR 2097887: Perf improvements for newtonsoft lottie  json reader
Perf improvements for newtonsoft lottie  json reader. This change eliminates the expensive ".load" operation for JArray and JObject. It appears that we never try to create a JArray or JObject from JTokens that do not have a derived type of JArray or JObject. Before this change the perf for Newtonsoft was worse than Windows.Data.json with a large lottie file such as DailyOffers/paint.json taking 0.015 minutes for the Wrt reader and about 0.03 minutes for the Newtonsoft reader. With this change the Netwtonsoft reader takes about 0.0083 minutes on the same file.